moved the photo upload and add to a service, to reduce lines in controller
